Moved modules required for plugins to `s4t-lightning-rod/lib/pluginapi`, made their imports in the plugins.examples absolute...
So a plugin developer is supposed to `npm install s4t-lightning-rod` and import the pluginapi stuff just like from a usual npm module.

But on the LR side we have to hook require in plugins in order to replace these absolute imports with actual paths for these files in a current setup.

See the `lib/process/plugin/patch-require.js` for more comments